+++ layeredconfig/etcdstore.py
@@ -3,7 +3,7 @@ from . import ConfigSource
 
 import requests
 
-class EtcdSource(ConfigSource):
+class EtcdStore(ConfigSource):
 
     def __init__(self, baseurl="http://127.0.0.1:4001/v2/",
                  **kwargs):
@@ -16,6 +16,7 @@ class EtcdSource(ConfigSource):
         ``etcd`` has no concept of typed values, so all data from this
         source are returned as strings.
         """
+        super(EtcdStore, self).__init__(**kwargs)
         if kwargs.get('source'):
             # subsection
             self.source = kwargs['source']
@@ -25,7 +26,10 @@ class EtcdSource(ConfigSource):
             self.sectionkey = "/"
         resp = requests.get(self.source + self.sectionkey)
         self.values = resp.json()['node']['nodes']
-        
+        self.dirtyvalues = {}
+        self.writable = kwargs.get("writable", True)
+        self.subsectioncache = {}
+
     def has(self, key):
         for child in self.values:
             if 'dir' not in child and self.sectionkey + key == child['key']:
@@ -37,12 +41,12 @@ class EtcdSource(ConfigSource):
             if self.sectionkey + key == child['key']:
                 return child['value']
         raise KeyError(key)
-            
+
     def keys(self):
         for child in self.values:
             if 'dir' not in child:
                 yield child['key'][len(self.sectionkey):]
-            
+
     def typed(self, key):
         return False # in etcd, all keys seem to be strings. Or can
                      # they be ints, bools and lists (JSON supported
@@ -52,22 +56,29 @@ class EtcdSource(ConfigSource):
         for child in self.values:
             if 'dir' in child:
                 yield child['key'][len(self.sectionkey):]
-        
+
     def subsection(self, key):
-        prefix = self.sectionkey
-        if not prefix.endswith("/"):
-            prefix += "/"
-        return EtcdSource(source=self.source, sectionkey=prefix+key+"/")
-    
+        if key not in self.subsectioncache:
+            prefix = self.sectionkey
+            if not prefix.endswith("/"):
+                prefix += "/"
+            self.subsectioncache[key] = EtcdStore(source=self.source,
+                                                  parent=self,
+                                                  sectionkey=prefix+key+"/")
+        return self.subsectioncache[key]
+
     def set(self, key=None, value=None):
         self.dirty = True
         if self.parent:
             self.parent.set()
-        if not self.dirtyvalues:
-            self.dirtyvalues = []
         if key and value:
             self.dirtyvalues[key] = value
-        
+
     def save(self):
         for k in self.dirtyvalues:
-            self.source.write(k, dirty[k])
+            requests.put(self.source+self.sectionkey+k,
+                         data={'value': str(self.dirtyvalues[k])})
+        self.dirtyvalues = {}
+        for subsection in self.subsections():
+            self.subsection(subsection).save()
+        self.dirty = False