+++ HISTORY.rst
@@ -8,7 +8,8 @@ History
 
 * Integration with argparse: The Commandline source now accepts an
   optional parse parameter, which should be a configured
-  argparse.ArgumentParser object. 
+  argparse.ArgumentParser object. Most features of argparse, such as
+  specifying the type of arguments, and automatic help text 
 
 0.1.0 (2014-11-03)
 ------------------